pyproject.toml
[tool.poetry]
name = "stag-erasmus"
version = "0.1.0"
description = ""
authors = ["yoptohlejepeta <peta.kotlan@seznam.cz>"]
readme = "README.md"
packages = [{include = "stag_erasmus"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.108.0"
requests = "^2.31.0"
pandas = "^2.1.4"
pydantic = "^2.5.3"
jinja2 = "^3.1.2"
unidecode = "^1.3.7"
uvicorn = "^0.25.0"
python-multipart = "^0.0.6"
redis = "^5.0.1"
aiohttp = "^3.9.3"
python-dotenv = "^1.0.1"
logger = "^1.4"
loguru = "^0.7.2"
tqdm = "^4.66.4"
schedule = "^1.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["ALL"]
exclude = ["*.ipynb", "__init__.py"]
ignore = [
    "FA102",
    "D211",
    "D213",
    "COM812",
    "ISC001",
    "BLE001",
    "D203",
    "B008",
    "PLR0913",
    "FBT001",
    "N805",
    "TD002",
    "TD003",
    "FIX002",
    "DTZ005",
    "RUF013",
    "FBT002",
    "S113",
    "PLC2401",
    ]
